deps/v8/test/inspector/debugger/set-async-call-stack-depth-expected.txt

Checks that we report not more then maxDepth call chains.

Running test: testPaused
Actual call chain length: 4
setAsyncCallStackDepth(maxDepth): 8
reported: 4

Actual call chain length: 4
setAsyncCallStackDepth(maxDepth): 4
reported: 4

Actual call chain length: 4
setAsyncCallStackDepth(maxDepth): 3
reported: 3

Actual call chain length: 4
setAsyncCallStackDepth(maxDepth): 0
reported: 0


Running test: testConsoleTrace
Actual call chain length: 4
setAsyncCallStackDepth(maxDepth): 8
reported: 4

Actual call chain length: 4
setAsyncCallStackDepth(maxDepth): 4
reported: 4

Actual call chain length: 4
setAsyncCallStackDepth(maxDepth): 3
reported: 3

Actual call chain length: 4
setAsyncCallStackDepth(maxDepth): 0
reported: 0
